+++ src/matterbridgeAccessoryPlatform.ts
@@ -49,6 +49,8 @@ export class MatterbridgeAccessoryPlatform {
 
   /**
    * This method must be overridden in the extended class.
+   * It is called when the platform is started.
+   * Use this method to create the MatterbridgeDevice and call this.registerDevice().
    * @param {string} [reason] - The reason for starting.
    * @throws {Error} - Throws an error if the method is not overridden.
    */
@@ -59,12 +61,31 @@ export class MatterbridgeAccessoryPlatform {
 
   /**
    * This method must be overridden in the extended class.
+   * It is called when the platform is shutting down.
+   * Use this method to clean up any resources.
    * @param {string} [reason] - The reason for shutting down.
    * @throws {Error} - Throws an error if the method is not overridden.
    */
   async onShutdown(reason?: string) {
     this.log.error('Plugins must override onShutdown.', reason);
     throw new Error('Plugins must override onShutdown.');
+  }
+
+  /**
+   * This method can be overridden in the extended class.
+   * It is called after matter server started.
+   */
+  async onMatterStarted() {
+    this.log.debug('The plugin can override onMatterStarted.');
+  }
+
+  /**
+   * This method can be overridden in the extended class.
+   * It is called after the platform has been commissioned.
+   * Use this method to perform any configuration of your devices.
+   */
+  async onConfigure() {
+    this.log.debug('The plugin can override onConfigure.');
   }
 
   /**